+++ mirage/configure/configure.py
@@ -69,7 +69,7 @@ class ConfigureWorkFlow(Workflow):
                 log("Miragefile is already exists!", withError = True)
                 raise FileExistsError
                 return
-        
+
         option_string = log("Additional option string", withInput = True)
 
         with open("Miragefile.addon", "w") as f: